pyproject.toml
[project]
name = "silent-binary-hardening-issues-extractor"
version = "0.1.0"
description = "自动化检测二进制防御机制缺失的Silent Bug提取工具"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "pyyaml>=6.0",
    "lxml>=4.9.0",
    "playwright>=1.52.0",
]

[project.scripts]
silent-extractor = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.sdist]
include = ["/src", "/config", "/README.md", "/pyproject.toml"]
